dom/security/test/csp/test_report_uri_missing_in_report_only_header.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=847081
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 847081</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=847081">Mozilla Bug 847081</a>
<p id="display"></p>
<div id="content" style="display: none"></div>
<iframe id="cspframe"></iframe>

<pre id="test">
<script class="testbody" type="text/javascript">
var stringBundleService = SpecialPowers.Cc["@mozilla.org/intl/stringbundle;1"]
                          .getService(SpecialPowers.Ci.nsIStringBundleService);
var localizer = stringBundleService.createBundle("chrome://global/locale/security/csp.properties");
var warningMsg = localizer.formatStringFromName("reportURInotInReportOnlyHeader", [window.location.origin]);

function cleanup() {
  SpecialPowers.postConsoleSentinel();
  SimpleTest.finish();
}

// Since Bug 1584993 we parse the CSP in the parent too, hence the
// same error message appears twice in the console. 
var recordConsoleMsgOnce = false;

SpecialPowers.registerConsoleListener(function ConsoleMsgListener(aMsg) {
  if (aMsg.message.indexOf(warningMsg) > -1) {
    if (recordConsoleMsgOnce) {
      return;
    }
    recordConsoleMsgOnce = true;

    ok(true, "report-uri not specified in Report-Only should throw a CSP warning.");
    SimpleTest.executeSoon(cleanup);
  }
  // Otherwise, if some other console message is present, we wait.
});


// set up and start testing
SimpleTest.waitForExplicitFinish();
document.getElementById('cspframe').src = 'file_report_uri_missing_in_report_only_header.html';
</script>
</pre>
</body>
</html>
